ENH: Addition of data files to download example data json files can be loaded in the python script. These new json files contain information about files that can be downloaded by the script at runtime. The url and image file names and IDs are not hardcoded in the script, but are saved in these external json files. LowRankImageDecomposition.py is a convenience script to convert XML files downloaded from a MIDAS server into a json file.
ENH: Multithreading slow operations using the paradigm used for the SimpleITK filters

The method downloading the data from a Midas server (URL and files to download are store in a JSON file) and
running the pyLAR algorithms are multithreaded using the method implemented in [1]. This allows the Slicer GUI
to stay responsive while one of these operation is performed. Since Slicer creashes if new data is loaded from
a thread that is not the main thread, the new thread only performs computation and file management operations.
Images computed or downloaded in the secondary thread are past to the main thread through a queue that loads the images
using a QTimer.