ENH: Addition of data files to download example data
json files can be loaded in the python script. These new json files contain information about files that can be downloaded by the script at runtime. The url and image file names and IDs are not hardcoded in the script, but are saved in these external json files.
LowRankImageDecomposition.py is a convenience script to convert XML files downloaded from a MIDAS server into a json file.