ENH: ANTS uses ITK from Slicer
When building ANTS with SuperBuild set to ON and running the executable from the ANTS package from within Slicer, runtime errors and segfaults happen. To avoid this issue, we use ITK from Slicer to build ANTS. Because ANTS only compiles with very recent versions on ITK, it only compiles if Slicer is very recent to (Slicer nightly > Slicer 4.5).