BUG: Trying to remove output directory even if it doesn't exist
When setting 'clean' to 'True' in the configuration file, the output directory is remove to make sure that computation is started from scratch. If the directory does not exist, the script should not try to remove it, or an error would be raised. To avoid this problem, we first check if the directory already exists.