ENH: Packaging process updated
Instead of installing external projects to know the location of their executables
(the location of the executables is unknown at configuration time on some configurations),
the installation is now done with CPACK_INSTALL_CMAKE_PROJECTS. ANTS was updated to take
the most of this process. It only compiles and packages the required tools.